+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ### [unreleased](https://github.com/speelbarrow/weensy.rs/blob/main)
+- Fix [verification of function signature by `entry` attribute macro](https://github.com/speelbarrow/)
 - Add/edit [badges in README.md](https://github.com/speelbarrow/weensy.rs/blob/main/README.md?plain=1#L1)
 
 ### [0.1.1](https://github.com/speelbarrow/weensy.rs/blob/v0.1.1)